added timeout for get, re-added import, and moved release_info to original location
