+++ mitreattack/diffStix/changelog_helper.py
@@ -22,7 +22,7 @@ from requests.adapters import HTTPAdapter, Retry
 from rich.progress import track
 from stix2 import Filter, MemoryStore
 from tqdm import tqdm
-import release_info
+from mitreattack import release_info
 
 # explanation of modification types to data objects for legend in layer files
 date = datetime.datetime.today()
@@ -542,22 +542,18 @@ class DiffStix(object):
             STIX MemoryStore object representing an ATT&CK domain.
         """
         error_message = f"Unable to successfully download ATT&CK STIX data from GitHub for {domain}. Please try again."
+        s = requests.Session()
+        retries = Retry(total=10, backoff_factor=0.3, status_forcelist=[500, 502, 503, 504])
+        s.mount("http", HTTPAdapter(max_retries=retries))
         try:
-            s = requests.Session()
-            retries = Retry(total=10, backoff_factor=0.3, status_forcelist=[500, 502, 503, 504])
-            s.mount("http", HTTPAdapter(max_retries=retries))
-            stix_response = s.get(f"https://raw.githubusercontent.com/mitre/cti/master/{domain}/{domain}.json")
+            stix_response = s.get(f"https://raw.githubusercontent.com/mitre/cti/master/{domain}/{domain}.json",
+                                  timeout=60)
             if stix_response.status_code != 200:
                 logger.error(error_message)
                 sys.exit(1)
-        except requests.exceptions.ContentDecodingError:
-            logger.error(error_message)
-            sys.exit(1)
-        except requests.exceptions.JSONDecodeError:
-            s2 = requests.Session()
-            retries = Retry(total=10, backoff_factor=0.1, status_forcelist=[500, 502, 503, 504])
-            s2.mount("http", HTTPAdapter(max_retries=retries))
-            stix_response = s2.get(f"https://raw.githubusercontent.com/mitre/cti/master/{domain}/{domain}.json")
+        except (requests.exceptions.ContentDecodingError, requests.exceptions.JSONDecodeError) as error:
+            stix_response = s.get(f"https://raw.githubusercontent.com/mitre/cti/master/{domain}/{domain}.json",
+                                  timeout=60)
             if stix_response.status_code != 200:
                 logger.error(error_message)
                 sys.exit(1)
